requirements.txt
google-adk>=1.2.1
a2a-sdk==0.2.8
httpx>=0.28.1
langchain-google-genai>=2.0.10
langgraph>=0.3.18
pydantic>=2.10.6
python-dotenv>=1.1.0
uvicorn>=0.34.2
nest-asyncio>=1.6.0
langchain-core
langchain-google-community==2.0.7
google-auth-oauthlib
langchain_mcp_adapters
neo4j
pysolr
